+++ cgv_clone/src/main/java/com/ceos22/cgv_clone/service/AuthService.java
@@ -1,9 +1,6 @@
 package com.ceos22.cgv_clone.service;
 
-import com.ceos22.cgv_clone.domain.dto.Member;
-import com.ceos22.cgv_clone.domain.dto.SignUpReq;
-import com.ceos22.cgv_clone.domain.member.MemberEntity;
-import com.ceos22.cgv_clone.repository.MemberRepository;
+import com.ceos22.cgv_clone.api.dto.SignUpRequest;
 import com.ceos22.cgv_clone.service.member.CreateMemberCommand;
 import com.ceos22.cgv_clone.service.member.MemberReader;
 import com.ceos22.cgv_clone.service.member.MemberSaver;
@@ -11,8 +8,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,12 +18,9 @@ public class AuthService {
     private final PasswordEncoder passwordEncoder; // BCryptPasswordEncoder
 
     @Transactional
-    public void signUp(SignUpReq req) {
+    public void signUp(SignUpRequest req) {
         // 이미 존재하는 loginId 체크
-        Optional<Member> member = Optional.ofNullable(memberReader.getByLoginIdOrThrow(req.loginId()));
-        if(member.isPresent()){
-            throw new IllegalArgumentException("member alerady ");
-        }
+        memberReader.getByLoginId(req.loginId());
 
         // 새 회원 생성 (비밀번호는 반드시 암호화)
         CreateMemberCommand m = new CreateMemberCommand(
